requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
mysql-connector==2.2.9
oandapyV20==0.6.3
requests==2.22.0
six==1.12.0
urllib3==1.25.3
